src/main/java/com/netcracker/odmg/secondgroup/zoxal/nc2_libraryCard/model/Record.java
package com.netcracker.odmg.secondgroup.zoxal.nc2_libraryCard.model;

public class Record {
	private int id;
	private String title;
	private String author;	
	private String obtainDate;
	private String returnDate;
	
	public void setId(int id) {
		this.id = id;
	}
	public void setTitle(String title) {
		this.title = title;
	}
	public void setAuthor(String author) {
		this.author = author;
	}
	public void setObtainDate(String obtainDate) {
		this.obtainDate = obtainDate;
	}
	public void setReturnDate(String returnDate) {
		this.returnDate = returnDate;
	}
	
	public int getId() {
		return id;
	}
	public String getTitle() {
		return title;
	}
	public String getAuthor() {
		return author;
	}
	public String getObtainDate() {
		return obtainDate;
	}
	public String getReturnDate() {
		return returnDate;
	}
	
	@Override
	public int hashCode() {
		final int prime = 31;
		int result = 1;
		result = prime * result + ((author == null) ? 0 : author.hashCode());
		result = prime * result + id;
		result = prime * result + ((obtainDate == null) ? 0 : obtainDate.hashCode());
		result = prime * result + ((returnDate == null) ? 0 : returnDate.hashCode());
		result = prime * result + ((title == null) ? 0 : title.hashCode());
		return result;
	}
	@Override
	public boolean equals(Object obj) {
		if (this == obj) {
			return true;
		}
		if (obj == null) {
			return false;
		}
		if (!(obj instanceof Record)) {
			return false;
		}
		Record other = (Record) obj;
		if (author == null) {
			if (other.author != null) {
				return false;
			}
		} else if (!author.equals(other.author)) {
			return false;
		}
		if (id != other.id) {
			return false;
		}
		if (obtainDate == null) {
			if (other.obtainDate != null) {
				return false;
			}
		} else if (!obtainDate.equals(other.obtainDate)) {
			return false;
		}
		if (returnDate == null) {
			if (other.returnDate != null) {
				return false;
			}
		} else if (!returnDate.equals(other.returnDate)) {
			return false;
		}
		if (title == null) {
			if (other.title != null) {
				return false;
			}
		} else if (!title.equals(other.title)) {
			return false;
		}
		return true;
	}
}
